+++ overlays/arcmenu.nix
@@ -1,0 +1,10 @@
+final: prev: {
+  gnomeExtensions = prev.gnomeExtensions // {
+    arc-menu = prev.gnomeExtensions.arc-menu.overrideAttrs (old: {
+      nativeBuildInputs = (old.nativeBuildInputs or [ ]) ++ [
+        final.buildPackages.glib
+        final.buildPackages.gitMinimal
+      ];
+    });
+  };
+}