+++ overlays/gom.nix
@@ -1,0 +1,5 @@
+final: prev: {
+  gom = prev.gom.overrideAttrs (old: {
+    nativeBuildInputs = (old.nativeBuildInputs or [ ]) ++ [ final.buildPackages.python3 ];
+  });
+}