+++ profiles/base/default.nix
@@ -1,0 +1,14 @@
+{ pkgs, lib, ... }:
+{
+  environment.systemPackages = with pkgs; [
+    curl
+    wget
+    htop
+    git
+    tmux
+    vim
+    btop
+    librewolf
+    ungoogled-chromium
+  ];
+}