+++ profiles/development/default.nix
@@ -1,0 +1,24 @@
+{
+  pkgs,
+  lib,
+  secrets,
+  ...
+}:
+let
+  inherit (lib) mkIf;
+in
+{
+  imports = [ ../base/default.nix ];
+
+  environment.systemPackages = with pkgs; [
+    docker-compose
+    docker
+    vscodium-fhs
+    opencode
+  ];
+
+  environment.variables = {
+    # GITHUB_TOKEN = mkIf (secrets ? githubToken) secrets.githubToken;
+    # NPM_TOKEN = mkIf (secrets ? npmToken) secrets.npmToken;
+  };
+}