+++ shell.nix
@@ -1,0 +1,10 @@
+{
+  pkgs ? import <nixpkgs> { },
+}:
+
+pkgs.mkShell {
+  packages = with pkgs; [
+    nixfmt-tree
+    nix-output-monitor
+  ];
+}