+++ variants/aarch64-luks-botany-bay.nix
@@ -1,0 +1,9 @@
+{
+  host = "Aarch64LUKS";
+  desktop = "gnome";
+  profiles = [ "botany-bay" ];
+  platform = "generic";
+  buildType = "none";
+  cross = false;
+  configName = "Aarch64LUKSBotanyBay";
+}