+++ variants/generic-aarch64-installer.nix
@@ -1,0 +1,5 @@
+{
+  host = "GenericAarch64";
+  platform = "generic";
+  buildType = "installer";
+}