+++ variants/pinebookpro-dev.nix
@@ -1,0 +1,10 @@
+{
+  host = "PineBookPro";
+  desktop = "gnome";
+  profiles = [
+    "base"
+    "development"
+  ];
+  platform = "rockchip";
+  buildType = "image";
+}