+++ variants/pinebookpro-plasma.nix
@@ -1,0 +1,6 @@
+{
+  host = "PineBookPro";
+  desktop = "plasma";
+  platform = "rockchip";
+  buildType = "image";
+}