+++ variants/pinetab2-gnome.nix
@@ -1,0 +1,6 @@
+{
+  host = "PineTab2";
+  desktop = "gnome";
+  platform = "rockchip";
+  buildType = "image";
+}