+++ variants/uboot.nix
@@ -1,0 +1,5 @@
+{
+  host = "PineBookPro";
+  platform = "rockchip";
+  buildType = "uboot";
+}