+++ variants/x86pc-plasma.nix
@@ -1,0 +1,6 @@
+{
+  host = "X86Pc";
+  desktop = "plasma";
+  platform = "x86";
+  buildType = "iso";
+}